Add files via upload
Refactor code and add macroquad user interface
